fix(budget): 예산 금액 필드 Int → Long 타입 변경 및 날짜 @NotNull 검증 추가

+++ src/main/kotlin/team/nongchun/hororog/domain/budget/dto/BudgetResponse.kt
@@ -7,8 +7,8 @@ import java.time.LocalDateTime
 data class BudgetResponse(
     val id: Long,
     val title: String,
-    val totalAmount: Int,
-    val usedAmount: Int,
+    val totalAmount: Long,
+    val usedAmount: Long,
     val startDate: LocalDate,
     val endDate: LocalDate,
     val createdAt: LocalDateTime,
